adding type annotations
