--- deploy-commands.js
@@ -1,24 +1,0 @@
-const fs = require("fs");
-const { REST } = require("@discordjs/rest");
-const { Routes } = require("discord-api-types/v9");
-const { clientId, guildIds, token } = require("./config.json");
-
-const commands = [];
-const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
-
-for (const file of commandFiles) {
-  const command = require(`./commands/${file}`);
-  commands.push(command.data.toJSON());
-}
-
-const rest = new REST({ version: "9" }).setToken(token);
-
-//register commands to multiple guilds if needed
-(async () => {
-  guildIds.forEach((guildId) => {
-    rest
-      .put(Routes.applicationGuildCommands(clientId, guildId), { body: commands })
-      .then(() => console.log("Successfully registered application commands."))
-      .catch(console.error);
-  });
-})();
