+++ deploy-commands.ts
@@ -1,0 +1,26 @@
+const fs = require("fs");
+const { REST } = require("@discordjs/rest");
+const { Routes } = require("discord-api-types/v9");
+const { clientId, guildIds, token } = require("./config.json");
+
+const commands: Array<any> = [];
+const commandFiles: Array<string> = fs.readdirSync('./commands')
+  .filter((file: string) => file.endsWith('.ts'));
+
+for (const file of commandFiles) {
+  const command = require(`./commands/${file}`);
+  commands.push(command.data.toJSON());
+}
+
+const rest = new REST({ version: "9" }).setToken(token);
+
+(async () => {
+  guildIds.forEach((guildId: string) => {
+    rest
+      .put(Routes.applicationGuildCommands(clientId, guildId), { body: commands })
+      .then(() => console.log("Successfully registered application commands."))
+      .catch(console.error);
+  });
+})();
+
+export { };
